backend/requirements.txt
eventlet==0.24.1
numpy
opencv-python-headless
flask==1.0.2
flask-cors==3.0.7
flask-login==0.4.1
flask-restplus==0.12.1
flask-mongoengine==0.9.5
cython
scikit-image
requests
google_images_download==2.5.0
watchdog==0.8.3
pytest==3.9.3
pytest-ordering==0.6
imantics==0.1.9
flask-socketio==3.3.2
celery==4.2.2
Shapely==1.7.0
scipy
Pillow
matplotlib
keras==2.1.1
h5py
imgaug
IPython[all]
jupyter
torch
torchvision
requests_toolbelt
albumentations>0.5
graphviz
tqdm
pyyaml
easydict
tensorboard
future
cffi
ninja